+++ Src/Library/Endpoint/Auxiliary/Group.cs
@@ -5,7 +5,7 @@ namespace FastEndpoints;
 /// <summary>
 /// common configuration for a group of endpoints can be specified by implementing this abstract class and calling <see cref="Configure(string, Action{EndpointDefinition})"/> in the constructor.
 /// </summary>
-public abstract class Group : IServiceResolver
+public abstract class Group : IServiceResolverBase
 {
     internal Action<EndpointDefinition> Action { get; set; }
 
@@ -14,8 +14,7 @@ public abstract class Group : IServiceResolver
     /// </summary>
     /// <param name="routePrefix">the route prefix for the group</param>
     /// <param name="ep">the configuration action to be performed on the <see cref="EndpointDefinition"/></param>
-    protected virtual void Configure(string routePrefix, Action<EndpointDefinition> ep)
-        => Action = RouteModifier(routePrefix) + ep;
+    protected virtual void Configure(string routePrefix, Action<EndpointDefinition> ep) => Action = RouteModifier(routePrefix) + ep;
 
     private static Action<EndpointDefinition> RouteModifier(string routePrefix) => e =>
     {
@@ -30,43 +29,16 @@ public abstract class Group : IServiceResolver
         }
     };
 
-    /// <summary>
-    /// try to resolve an instance for the given type from the dependency injection container. will return null if unresolvable.
-    /// </summary>
-    /// <typeparam name="TService">the type of the service to resolve</typeparam>
-    public TService? TryResolve<TService>() where TService : class
-        => IServiceResolver.RootServiceProvider.GetService<TService>();
-    /// <summary>
-    /// try to resolve an instance for the given type from the dependency injection container. will return null if unresolvable.
-    /// </summary>
-    /// <param name="typeOfService">the type of the service to resolve</param>
-    public object? TryResolve(Type typeOfService)
-        => IServiceResolver.RootServiceProvider.GetService(typeOfService);
-    /// <summary>
-    /// resolve an instance for the given type from the dependency injection container. will throw if unresolvable.
-    /// </summary>
-    /// <typeparam name="TService">the type of the service to resolve</typeparam>
-    /// <exception cref="InvalidOperationException">Thrown if requested service cannot be resolved</exception>
-    public TService Resolve<TService>() where TService : class
-        => IServiceResolver.RootServiceProvider.GetRequiredService<TService>();
-    /// <summary>
-    /// resolve an instance for the given type from the dependency injection container. will throw if unresolvable.
-    /// </summary>
-    /// <param name="typeOfService">the type of the service to resolve</param>
-    /// <exception cref="InvalidOperationException">Thrown if requested service cannot be resolved</exception>
-    public object Resolve(Type typeOfService)
-        => IServiceResolver.RootServiceProvider.GetRequiredService(typeOfService);
-    /// <summary>
-    /// if you'd like to resolve scoped or transient services from the DI container, obtain a service scope from this method and dispose the scope when the work is complete.
-    ///<para>
-    /// <code>
-    /// using var scope = CreateScope();
-    /// var scopedService = scope.ServiceProvider.GetService(...);
-    /// </code>
-    /// </para>
-    /// </summary>
-    public IServiceScope CreateScope()
-        => IServiceResolver.RootServiceProvider.CreateScope();
+    ///<inheritdoc/>
+    public TService? TryResolve<TService>() where TService : class => Config.ServiceResolver.TryResolve<TService>();
+    ///<inheritdoc/>
+    public object? TryResolve(Type typeOfService) => Config.ServiceResolver.TryResolve(typeOfService);
+    ///<inheritdoc/>
+    public TService Resolve<TService>() where TService : class => Config.ServiceResolver.Resolve<TService>();
+    ///<inheritdoc/>
+    public object Resolve(Type typeOfService) => Config.ServiceResolver.Resolve(typeOfService);
+    ///<inheritdoc/>
+    public IServiceScope CreateScope() => Config.ServiceResolver.CreateScope();
 }
 
 /// <summary>