+++ Src/Library/Endpoint/Endpoint.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
-using static FastEndpoints.Config;
 
 namespace FastEndpoints;
 
@@ -35,7 +34,7 @@ public abstract class EndpointWithMapper<TRequest, TMapper> : Endpoint<TRequest,
 /// </summary>
 /// <typeparam name="TRequest">the type of the request dto</typeparam>
 /// <typeparam name="TResponse">the type of the response dto</typeparam>
-public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint, IServiceResolver where TRequest : notnull, new()
+public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint, IServiceResolverBase where TRequest : notnull, new()
 {
     internal async override Task ExecAsync(CancellationToken ct)
     {
@@ -44,12 +43,12 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint, ISer
         try
         {
             var binder = (IRequestBinder<TRequest>)
-                 (Definition.RequestBinder ??= HttpContext.RequestServices.GetRequiredService(typeof(IRequestBinder<TRequest>)));
+                 (Definition.RequestBinder ??= FastEndpoints.Config.ServiceResolver.Resolve(typeof(IRequestBinder<TRequest>)));
 
             var binderCtx = new BinderContext(HttpContext, ValidationFailures, Definition.SerializerContext, Definition.DontBindFormData);
             req = await binder.BindAsync(binderCtx, ct);
 
-            BndOpts.Modifier?.Invoke(req, tRequest, binderCtx, ct);
+            FastEndpoints.Config.BndOpts.Modifier?.Invoke(req, tRequest, binderCtx, ct);
 
             OnBeforeValidate(req);
             await OnBeforeValidateAsync(req, ct);
@@ -90,7 +89,7 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint, ISer
             await OnValidationFailedAsync(ct);
 
             if (!Definition.DoNotCatchExceptions)
-                await SendErrorsAsync(ErrOpts.StatusCode, ct);
+                await SendErrorsAsync(FastEndpoints.Config.ErrOpts.StatusCode, ct);
             else
                 throw;
         }
@@ -116,38 +115,16 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint, ISer
     [NotImplemented]
     public virtual Task<TResponse> ExecuteAsync(TRequest req, CancellationToken ct) => throw new NotImplementedException();
 
-    /// <summary>
-    /// try to resolve an instance for the given type from the dependency injection container. will return null if unresolvable.
-    /// </summary>
-    /// <typeparam name="TService">the type of the service to resolve</typeparam>
-    public TService? TryResolve<TService>() where TService : class => HttpContext.RequestServices.GetService<TService>();
-    /// <summary>
-    /// try to resolve an instance for the given type from the dependency injection container. will return null if unresolvable.
-    /// </summary>
-    /// <param name="typeOfService">the type of the service to resolve</param>
-    public object? TryResolve(Type typeOfService) => HttpContext.RequestServices.GetService(typeOfService);
-    /// <summary>
-    /// resolve an instance for the given type from the dependency injection container. will throw if unresolvable.
-    /// </summary>
-    /// <typeparam name="TService">the type of the service to resolve</typeparam>
-    /// <exception cref="InvalidOperationException">Thrown if requested service cannot be resolved</exception>
-    public TService Resolve<TService>() where TService : class => HttpContext.RequestServices.GetRequiredService<TService>();
-    /// <summary>
-    /// resolve an instance for the given type from the dependency injection container. will throw if unresolvable.
-    /// </summary>
-    /// <param name="typeOfService">the type of the service to resolve</param>
-    /// <exception cref="InvalidOperationException">Thrown if requested service cannot be resolved</exception>
-    public object Resolve(Type typeOfService) => HttpContext.RequestServices.GetRequiredService(typeOfService);
-    /// <summary>
-    /// if you'd like to resolve scoped or transient services from the DI container, obtain a service scope from this method and dispose the scope when the work is complete.
-    ///<para>
-    /// <code>
-    /// using var scope = CreateScope();
-    /// var scopedService = scope.ServiceProvider.GetService(...);
-    /// </code>
-    /// </para>
-    /// </summary>
-    public IServiceScope CreateScope() => HttpContext.RequestServices.CreateScope();
+    /// <inheritdoc/>
+    public TService? TryResolve<TService>() where TService : class => FastEndpoints.Config.ServiceResolver.TryResolve<TService>();
+    /// <inheritdoc/>
+    public object? TryResolve(Type typeOfService) => FastEndpoints.Config.ServiceResolver.TryResolve(typeOfService);
+    ///<inheritdoc/>
+    public TService Resolve<TService>() where TService : class => FastEndpoints.Config.ServiceResolver.Resolve<TService>();
+    ///<inheritdoc/>
+    public object Resolve(Type typeOfService) => FastEndpoints.Config.ServiceResolver.Resolve(typeOfService);
+    ///<inheritdoc/>
+    public IServiceScope CreateScope() => FastEndpoints.Config.ServiceResolver.CreateScope();
 
     /// <summary>
     /// publish the given model/dto to all the subscribers of the event notification
@@ -160,7 +137,7 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint, ISer
     /// Mode.WaitForAny returns a Task that will complete when any of the subscribers complete their work.
     /// Mode.WaitForAll return a Task that will complete only when all of the subscribers complete their work.</returns>
     public Task PublishAsync<TEvent>(TEvent eventModel, Mode waitMode = Mode.WaitForAll, CancellationToken cancellation = default) where TEvent : class
-        => HttpContext.RequestServices.GetRequiredService<Event<TEvent>>().PublishAsync(eventModel, waitMode, cancellation);
+        => FastEndpoints.Config.ServiceResolver.Resolve<Event<TEvent>>().PublishAsync(eventModel, waitMode, cancellation);
 
     /// <summary>
     /// get the value of a given route parameter by specifying the resulting type and param name.