+++ Src/Library/Extensions/ReflectionExtensions.cs
@@ -200,4 +200,38 @@ internal static class ReflectionExtensions
             }
         }
     }
+
+    internal static Func<object, CancellationToken, Task<TResult>> HandlerExecutor<TResult>(this Type tHandler, Type tCommand, object handler)
+    {
+        //Task<TResult> ExecuteAsync((TCommand)cmd, ct);
+
+        var instance = Expression.Constant(handler);
+        var execMethod = tHandler.GetMethod("ExecuteAsync", BindingFlags.Instance | BindingFlags.Public | BindingFlags.FlattenHierarchy)!;
+        var cmdParam = Expression.Parameter(Types.Object, "cmd");
+        var ctParam = Expression.Parameter(typeof(CancellationToken), "ct");
+        var methodCall = Expression.Call(instance, execMethod, Expression.Convert(cmdParam, tCommand), ctParam);
+
+        return Expression.Lambda<Func<object, CancellationToken, Task<TResult>>>(
+            methodCall,
+            cmdParam,
+            ctParam
+        ).Compile();
+    }
+
+    internal static Func<object, CancellationToken, Task> HandlerExecutor(this Type tHandler, Type tCommand, object handler)
+    {
+        //Task ExecuteAsync((TCommand)cmd, ct);
+
+        var instance = Expression.Constant(handler);
+        var execMethod = tHandler.GetMethod("ExecuteAsync", BindingFlags.Instance | BindingFlags.Public | BindingFlags.FlattenHierarchy)!;
+        var cmdParam = Expression.Parameter(Types.Object, "cmd");
+        var ctParam = Expression.Parameter(typeof(CancellationToken), "ct");
+        var methodCall = Expression.Call(instance, execMethod, Expression.Convert(cmdParam, tCommand), ctParam);
+
+        return Expression.Lambda<Func<object, CancellationToken, Task>>(
+            methodCall,
+            cmdParam,
+            ctParam
+        ).Compile();
+    }
 }