+++ Src/Library/Interfaces/ICommand.cs
@@ -1,0 +1,12 @@
+﻿namespace FastEndpoints;
+
+/// <summary>
+/// interface for a command that does not return anything
+/// </summary>
+public interface ICommand { }
+
+/// <summary>
+/// interface for a command that returns a result
+/// </summary>
+/// <typeparam name="TResult">the type of the result that will be returned from the handler [<see cref="FastCommandHandler{TCommand}"/>] of this command</typeparam>
+public interface ICommand<out TResult> { }