+++ Web/[Features]/TestCases/EventHandlingTest/NewItemAddedToStock.cs
@@ -19,7 +19,7 @@ public class NotifyCustomers : FastEventHandler<NewItemAddedToStock>
     }
 }
 
-public class UpdateInventoryLevel : FastEndpoints.FastEventHandler<NewItemAddedToStock>
+public class UpdateInventoryLevel : FastEventHandler<NewItemAddedToStock>
 {
     public override Task HandleAsync(NewItemAddedToStock eventModel, CancellationToken ct)
     {