+++ Src/Library/changelog.md
@@ -1,12 +1,10 @@
-# DOCS TO DO:
-- since mappers are singletons, any deps injected in ctor also become singletons. even deps registered as scoped/transient.
-
 ### NEW
 - dependency resolving support for endpoint `Configure()` method
 - type safety for the shortcut http verb methods such as `Get()`, `Post()`, etc. [#info](https://fast-endpoints.com/docs/misc-conveniences#shorthand-route-configuration)
 - custom value parser registration at startup for any given type #285
 - specify whether to execute global pre/post processors before or after endpoint level processors #291
 - `[DontInject]` attribute for preventing property injection of endpoint properties
+- command bus pattern request/response messaging (independant of endpoints) #294
 
 ### IMPROVEMENTS
 - remove read-only properties from swagger request body #283